remove recreate command
